// src/types/relatorio/rporcamentaria20xe.ts
export type RpOrcamentaria20XEType = {
  codigo: number;
  ano: number;
  ugExec: number;
  nomeUgExec: string;
  natDespesa: number;
  pi: string;
  nomePi: string;
  neCcor: string;
  favorecido: string;
  nomeFavorecido: string;
  neDescricao: string;
  dataNe: string;
  rpnpLiquidar: number;
  rpnpPagar: number;
};